requirements.txt
discord.py==1.3.4
PyYAML==5.3.1
